requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
Pillow==11.0.0
opencv-python==4.10.0.84
aiosqlite==0.20.0
pydantic>=2.4.1,<2.10
pydantic-settings==2.6.1
openai==1.55.3
numpy==1.26.4
